+++ CHANGELOG.md
@@ -6,6 +6,8 @@
 - added ignore options to intel.console()
 - added chalk.enabled when colorize is used
 - added padding and truncation to printf
+- added Logger#handleExceptions to catch uncaught exceptions
+- added stack traces when an exception is logged
 - changed datetime to strftime, adds `%L` to output milliseconds
 
 ## v0.2.0 - 2013-10-04